- Reverted currentPlanes variable back to uint32. Desired affect of removing   casting of original commit is maintained.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@31399 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csplugincommon/opengl/shaderplugin.h
@@ -116,7 +116,7 @@ namespace CS
       class CS_CSPLUGINCOMMON_GL_EXPORT ClipPlanes
       {
         size_t maxPlanes;
-        size_t currentPlanes;
+        uint32 currentPlanes;
         
         CS::ShaderVarStringID svObjectToWorldInv;
         CS::ShaderVarStringID svWorldToCamera;
